backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]>=3.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
numpy>=1.26.4
pandas>=2.2.3
scikit-learn>=1.5.2
joblib>=1.4.2
scipy>=1.14.1
httpx>=0.27.2
google-generativeai>=0.8.0